lambda-functions/code-generator/requirements.txt
requests
boto3
urllib3
pybase64
botocore
openai
aiohttp
pyyaml